requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
pydantic==2.5.2
python-dotenv==1.0.0 